+++ src/Nova/Role.php
@@ -12,6 +12,12 @@ use Sereny\NovaPermissions\Models\Role as RoleModel;
 
 class Role extends Resource
 {
+    /**
+     * Indicates if the resource should be displayed in the sidebar.
+     *
+     * @var bool
+     */
+    public static $displayInNavigation = false;
 
     /**
      * The list of field name that should be hidden